% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/biogenic_substrates.R
\name{biogenic_substrates}
\alias{biogenic_substrates}
\title{Biogenic substrates}
\format{
A \link[tibble:tibble-package]{tibble} of three variables:
\describe{
\item{\code{bs_id}}{Biogenic substrate identifier. This is an identifier assigned
by the authors of this package.}
\item{\code{bs_name}}{Biogenic substrate name.}
\item{\code{bs_level}}{Biogenic substrate level.}
}
}
\source{
\url{https://archimer.ifremer.fr/doc/00736/84820/}
}
\description{
A data set of biogenic substrates as defined in Table 5 of the EMODnet Seabed
Habitats Data Product (2019).
}
\examples{
biogenic_substrates

}
\keyword{datasets}
