% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_manifest.R
\name{create_manifest}
\alias{create_manifest}
\title{Create a manifest of downloaded files}
\usage{
create_manifest(path, ignore = "MANIFEST")
}
\arguments{
\item{path}{A path to the directory containing the downloaded archives.}

\item{ignore}{A character vector of file names to be ignored in the
generation of the manifest.}
}
\value{
A \link[tibble:tibble-package]{tibble} of four columns:
\describe{
\item{\code{file}}{File name.}
\item{\code{date}}{Download date.}
\item{\code{size}}{File size.}
\item{\code{md5sum}}{MD5 checksum.}
}
}
\description{
\code{\link[=create_manifest]{create_manifest()}} creates a \link[tibble:tibble-package]{tibble} that lists
downloaded files, along with download date, file size and respective MD5
checksums.
}
