% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/download.R
\name{download_sources}
\alias{download_sources}
\title{Download sources}
\usage{
download_sources(
  path,
  sources = read_sources(),
  overwrite = FALSE,
  progress = TRUE,
  timeout = 60
)
}
\arguments{
\item{path}{A path to a directory where the files are to be saved.}

\item{sources}{A list or data frame containing (at least) the elements \code{file}
and \code{url}. The \code{file} should contain the name of the destination file on
disk, and \code{url} the source URL. By default, it takes the output of
\code{\link[=read_sources]{read_sources()}}.}

\item{overwrite}{Whether to overwrite existing files.}

\item{progress}{Whether to print download progress.}

\item{timeout}{Timeout in seconds.}
}
\value{
A \link[tibble:tibble-package]{tibble} providing a log on the downloaded
files.
}
\description{
\code{download_sources()} downloads data sources.
}
