% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/identify_eunis_biogenic_habitats.R
\name{eunis_biogenic_habitats}
\alias{eunis_biogenic_habitats}
\title{EUNIS biogenic habitats}
\source{
\url{https://archimer.ifremer.fr/doc/00736/84820/}
}
\usage{
eunis_biogenic_habitats(cs = c("EUNIS_M_2019", "EUNIS_M_2022"))
}
\arguments{
\item{cs}{EUNIS classification version, either \code{"EUNIS_M_2019"} or
\code{"EUNIS_M_2022"}. Default is \code{"EUNIS_M_2019"}.}
}
\value{
A subset \link[tibble:tibble-package]{tibble} of the
\link[eunis.habitats:eunis_habitats]{eunis_habitats} matching the criteria
indicated above.
}
\description{
\code{\link[=eunis_biogenic_habitats]{eunis_biogenic_habitats()}} returns a data set of EUNIS biogenic habitats
following the criteria indicated in Section 2.1 of the EMODnet Seabed
Habitats Data Product (2019):
\itemize{
\item Exclusion of those classes from the littoral zone --- that is any
class in which the second value in the alphanumeric habitat code is "A". This
is because EUSeaMap is only concerned with sublittoral and deep-sea zones.
\item Kept only those classes where "biogenic habitat" is the given substrate
type in the level 2 parent class. That is any class in which the third value
in the alphanumeric habitat code is "2".
}

TODO: Remaining criteria remain to be implemented.
}
\examples{
# By default EUNIS Marine Habitats are returned
eunis_biogenic_habitats()

# Alternatively request EUNIS Marine 2022
eunis_biogenic_habitats("EUNIS_M_2022")

}
