% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/EUSeaMap-package.R
\docType{package}
\name{EUSeaMap-package}
\alias{EUSeaMap}
\alias{EUSeaMap-package}
\title{EUSeaMap: EMODnet Broad-Scale Seabed Habitat Map for Europe}
\description{
The EMODnet broad-scale seabed habitat map for Europe is a comprehensive, free and ready-to-use broad-scale map of physical habitats, harmonising mapping procedures and fostering a common understanding among seabed mappers in Europe. The map is also known as EUSeaMap. This package provides a suite of utilities to generate the EUSeaMap.
}
\author{
\strong{Maintainer}: Ramiro Magno \email{rmagno@pattern.institute} (\href{https://orcid.org/0000-0001-5226-3441}{ORCID})

Authors:
\itemize{
  \item Pedro Monteiro \email{pmontei@ualg.pt} (\href{https://orcid.org/0000-0003-1446-8256}{ORCID})
  \item Luís Bentes \email{lbentes@ualg.pt} (\href{https://orcid.org/0000-0001-6884-2886}{ORCID})
}

Other contributors:
\itemize{
  \item Pattern Institute [copyright holder, funder]
  \item CCMAR [copyright holder, funder]
}

}
\keyword{internal}
