% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/list_map_files.R
\name{ls_map_files}
\alias{ls_map_files}
\title{List map files in archives}
\usage{
ls_map_files(
  path,
  pattern = "+\\\\.zip$",
  exclusions = "EMODnetSBH_Confidence.xlsx"
)
}
\arguments{
\item{path}{A path to a directory containing archives with map files.}

\item{pattern}{A pattern to filter archive file names.}

\item{exclusions}{A character vector of file names not to be included in the
listing.}
}
\value{
A \link[tibble:tibble-package]{tibble} of seven variables comprising
the map files contained in the archives, one row per \code{file}:
\describe{
\item{ar_path}{Full path to the archive file.}
\item{ar}{Archive file name.}
\item{gui}{Global unique identifier.}
\item{file}{File name.}
\item{file_type}{File type.}
\item{size}{File size in bytes.}
\item{date}{File creation date.}
}
}
\description{
\code{ls_map_files()} lists the map files contained within archives.
}
\details{
This function loops over all archives in \code{path} whose file name matches the
\code{pattern} and lists files within each archive.
}
