% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_dbf.R
\name{read_dbf}
\alias{read_dbf}
\title{Read a dbf file}
\usage{
read_dbf(dbf_file, ar = NULL)
}
\arguments{
\item{dbf_file}{A path to a dbf file.}

\item{ar}{A path to an archive zip file if the dbf is contained in said zip
file.}
}
\value{
A \link[tibble:tibble-package]{tibble} of data from the DBF file.
}
\description{
\code{\link[=read_dbf]{read_dbf()}} reads a dBase IV format file.
}
