% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_sources_list.R
\name{read_sources}
\alias{read_sources}
\title{Read data sources}
\usage{
read_sources(
  file = system.file("EUSeaMap_phase5_rawdata.yml", package = "EUSeaMap")
)
}
\arguments{
\item{file}{A path to a YAML file with data sources. By default it reads
the most recent data sources' list shipped with \code{{EUSeaMap}}.}
}
\value{
A \link[tibble:tibble-package]{tibble} of four variables:
\describe{
\item{\code{file}}{Data source file name.}
\item{\code{desc}}{Description of the data source.}
\item{\code{cs}}{Marine habitat classification system.}
\item{\code{url}}{The URL to the data source location.}
}
}
\description{
\code{read_sources()} reads a table of data sources.
}
\examples{
read_sources()

}
